+++ src/main/java/kristiania/no/jdbc/options/Option.java
@@ -3,12 +3,12 @@ package kristiania.no.jdbc.options;
 public class Option {
     long id;
     String optionName;
-    int questionId;
+    long questionId;
 
     public Option() {
     }
 
-    public Option(String optionName, int questionId) {
+    public Option(String optionName, long questionId) {
         this.optionName = optionName;
         this.questionId = questionId;
     }
@@ -21,11 +21,11 @@ public class Option {
         this.id = id;
     }
 
-    public int getQuestionId() {
+    public long getQuestionId() {
         return questionId;
     }
 
-    public void setQuestionId(int questionId) {
+    public void setQuestionId(long questionId) {
         this.questionId = questionId;
     }
